# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Shapes and Categories API",
    "version": "1.0.0",
    "description": "API documentation for managing categories and shapes."
  },
  "paths": {
...
    "/shapes": {
      "get": {
        "summary": "Get all shapes",
        "tags": [
          "Shapes"
        ],
        "parameters": [
          {
            "name": "filters",
            "in": "query",
            "description": "Dynamic filters for querying shapes. Use dot notation to specify filters (e.g., `filters[field][$operator]=value`).The following fields are supported for filtering:\n\n **`name`**, **`type`**, **`author`**, **`tags`**, **`category`**, **`version`**",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "examples": {
              "Example 1": {
                "value": {
                  "filters[name][$eq]": "MyShapeName4",
                  "filters[type][$ne]": "3D",
                  "filters[category][$in]": "676d21530b21683093603bdb"
                }
              },
              "Example 2": {
                "value": {
                  "filters[author][$eq]": "authorId123",
                  "filters[tags][$in]": "design"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of shapes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shape"
                      },
                      "description": "List of shapes"
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total number of shapes"
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "_id": "676e9af6c73809d1a6f04f02",
                      "createdAt": "2024-12-27T12:17:58.424Z",
                      "updatedAt": "2024-12-27T12:17:58.424Z",
                      "status": "active",
                      "name": "circle 2*2",
                      "type": "2D",
                      "attachTo": "categoryId latest",
                      "svgFile": "my-shape.svg",
                      "svgContent": "<svg>...</svg>",
                      "version": "1.0.1",
                      "category": "676e9af6c73809d1a6f04f02",
                      "metadata": {
                        "description": "A detailed description of the shape",
                        "applicationTypes": [
                          "website",
                          "tab"
                        ],
                        "customProperties": {
                          "name": "custom data"
                        },
                        "dependencies": [
                          {
                            "shapeId": "shapeId1",
                            "version": "1.0.0"
                          }
                        ]
                      },
                      "tags": [
                        "2d layer",
                        "component specific"
                      ],
                      "author": "authorId123",
                      "diagram_components": {
                        "component1": {
                          "type": "visual",
                          "position": "top-left"
                        }
                      },
                      "attachment_points": {
                        "point1": {
                          "coordinates": [
                            10,
                            20
                          ],
                          "type": "connection"
                        }
                      }
                    }
                  ],
                  "total": 1
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new shape",
        "tags": [
          "Shapes"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Shape"
              },
              "example": {
                "name": "layer shape",
                "type": "2D",
                "attachTo": "categoryId latest",
                "svgFile": "my-shape.svg",
                "svgContent": "<svg>...</svg>",
                "version": "1.0.1",
                "category": "676e9a0dc6b16c1303c76b7c",
                "metadata": {
                  "description": "A detailed description of the shape",
                  "applicationTypes": [
                    "website",
                    "tab"
                  ],
                  "customProperties": {
                    "name": "custom data"
                  },
                  "dependencies": [
                    {
                      "shapeId": "shapeId1",
                      "version": "1.0.0"
                    }
                  ]
                },
                "tags": [
                  "2d layer",
                  "component specific"
                ],
                "author": "authorId123",
                "diagram_components": [
                  {
                    "id": "shape-1736157286702-iynkldxxt",
                    "shape": "server-L",
                    "position": "top",
                    "source": "shape",
                    "relativeToId": null,
                    "attached2DShapes": [],
                    "attachmentPoints": [
                      {
                        "name": "attach-front-right",
                        "x": 226.2,
                        "y": 159.39999
                      },
                      {
                        "name": "attach-front-left",
                        "x": 63.700001,
                        "y": 159.39999
                      },
                      {
                        "name": "attach-back-left",
                        "x": 63.799999,
                        "y": 65.599998
                      },
                      {
                        "name": "attach-top",
                        "x": 144.89999,
                        "y": 76.400002
                      },
                      {
                        "name": "attach-back-right",
                        "x": 226.2,
                        "y": 65.599998
                      },
                      {
                        "name": "attach-bottom",
                        "x": 145,
                        "y": 148.60001
                      }
                    ]
                  },
                  {
                    "id": "shape-1736157286703-502hkdwrk",
                    "shape": "database-L",
                    "position": "top",
                    "source": "shape",
                    "relativeToId": "shape-1736157286702-iynkldxxt",
                    "attached2DShapes": [
                      {
                        "name": "react logo-L",
                        "attachedTo": "top"
                      }
                    ],
                    "attachmentPoints": [
                      {
                        "name": "attach-top",
                        "x": 106,
                        "y": 54
                      },
                      {
                        "name": "attach-bottom",
                        "x": 106,
                        "y": 126.2
                      },
                      {
                        "name": "attach-front-right",
                        "x": 187.2,
                        "y": 137
                      },
                      {
                        "name": "attach-back-right",
                        "x": 187.2,
                        "y": 43.200001
                      },
                      {
                        "name": "attach-front-left",
                        "x": 24.700001,
                        "y": 137
                      },
                      {
                        "name": "attach-back-left",
                        "x": 24.700001,
                        "y": 43.200001
                      }
                    ]
                  }
                ],
                "attachment_points": [
                  {
                    "name": "attach-top",
                    "x": 645,
                    "y": 504
                  },
                  {
                    "name": "attach-front-right",
                    "x": 726,
                    "y": 659
                  },
                  {
                    "name": "attach-front-left",
                    "x": 564,
                    "y": 659
                  },
                  {
                    "name": "attach-back-left",
                    "x": 564,
                    "y": 566
                  },
                  {
                    "name": "attach-back-right",
                    "x": 726,
                    "y": 566
                  },
                  {
                    "name": "attach-bottom",
                    "x": 645,
                    "y": 649
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Shape created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Shape"
                },
                "example": {
                  "_id": "676e9af6c73809d1a6f04f02",
                  "createdAt": "2024-12-27T12:17:58.424Z",
                  "updatedAt": "2024-12-27T12:17:58.424Z",
                  "status": "active",
                  "name": "2d layer latest",
                  "type": "2D",
                  "attachTo": "categoryId latest",
                  "svgFile": "my-shape.svg",
                  "svgContent": "<svg>...</svg>",
                  "version": "1.0.1",
                  "category": "676e9a0dc6b16c1303c76b7c",
                  "metadata": {
                    "description": "A detailed description of the shape",
                    "applicationTypes": [
                      "website",
                      "tab"
                    ],
                    "customProperties": {
                      "name": "custom data"
                    },
                    "dependencies": [
                      {
                        "shapeId": "shapeId1",
                        "version": "1.0.0"
                      }
                    ]
                  },
                  "tags": [
                    "2d layer",
                    "component specific"
                  ],
                  "author": "authorId123",
                  "diagram_components": [
                    {
                      "id": "shape-1736157286702-iynkldxxt",
                      "shape": "server-L",
                      "position": "top",
                      "source": "shape",
                      "relativeToId": null,
                      "attached2DShapes": [],
                      "attachmentPoints": [
                        {
                          "name": "attach-front-right",
                          "x": 226.2,
                          "y": 159.39999
                        },
                        {
                          "name": "attach-front-left",
                          "x": 63.700001,
                          "y": 159.39999
                        },
                        {
                          "name": "attach-back-left",
                          "x": 63.799999,
                          "y": 65.599998
                        },
                        {
                          "name": "attach-top",
                          "x": 144.89999,
                          "y": 76.400002
                        },
                        {
                          "name": "attach-back-right",
                          "x": 226.2,
                          "y": 65.599998
                        },
                        {
                          "name": "attach-bottom",
                          "x": 145,
                          "y": 148.60001
                        }
                      ]
                    },
                    {
                      "id": "shape-1736157286703-502hkdwrk",
                      "shape": "database-L",
                      "position": "top",
                      "source": "shape",
                      "relativeToId": "shape-1736157286702-iynkldxxt",
                      "attached2DShapes": [
                        {
                          "name": "react logo-L",
                          "attachedTo": "top"
                        }
                      ],
                      "attachmentPoints": [
                        {
                          "name": "attach-top",
                          "x": 106,
                          "y": 54
                        },
                        {
                          "name": "attach-bottom",
                          "x": 106,
                          "y": 126.2
                        },
                        {
                          "name": "attach-front-right",
                          "x": 187.2,
                          "y": 137
                        },
                        {
                          "name": "attach-back-right",
                          "x": 187.2,
                          "y": 43.200001
                        },
                        {
                          "name": "attach-front-left",
                          "x": 24.700001,
                          "y": 137
                        },
                        {
                          "name": "attach-back-left",
                          "x": 24.700001,
                          "y": 43.200001
                        }
                      ]
                    }
                  ],
                  "attachment_points": [
                    {
                      "name": "attach-top",
                      "x": 645,
                      "y": 504
                    },
                    {
                      "name": "attach-front-right",
                      "x": 726,
                      "y": 659
                    },
                    {
                      "name": "attach-front-left",
                      "x": 564,
                      "y": 659
                    },
                    {
                      "name": "attach-back-left",
                      "x": 564,
                      "y": 566
                    },
                    {
                      "name": "attach-back-right",
                      "x": 726,
                      "y": 566
                    },
                    {
                      "name": "attach-bottom",
                      "x": 645,
                      "y": 649
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/shapes/search/{searchText}": {
      "get": {
        "summary": "Search for parent shapes",
        "tags": [
          "Shapes"
        ],
        "description": "This endpoint allows you to search for parent shapes based on dynamic filters. Use dot notation to specify filters (e.g., `filters[field][$operator]=value`). The following fields are supported for filtering:\n\n **`name`**, **`type`**, **`author`**, **`tags`**, **`category`**, **`version`**",
        "parameters": [
          {
            "name": "searchText",
            "in": "path",
            "required": true,
            "description": "text to be search ",
            "schema": {
              "type": "string",
              "example": "3 level child"
            }
          },
          {
            "name": "filters",
            "in": "query",
            "description": "Dynamic filters for querying shapes. You can use dot notation to specify filters. The following fields are supported for filtering:\n\n **`name`**, **`type`**, **`author`**, **`tags`**, **`category`**, **`version`**",
            "required": false,
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "string"
              }
            },
            "examples": {
              "Example 1": {
                "value": {
                  "filters[type][$eq]": "2D"
                }
              },
              "Example 2": {
                "value": {
                  "filters[name][$eq]": "MyShapeName",
                  "filters[category][$in]": "676d21530b21683093603bdb"
                }
              },
              "Example 3": {
                "value": {
                  "filters[author][$eq]": "authorId123",
                  "filters[tags][$in]": "design"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of parent shapes matching the filters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shape"
                      },
                      "description": "List of parent shapes"
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total number of shapes matching the filter"
                    }
                  }
                },
                "example": {
                  "data": [
                    {
                      "_id": "676e9af6c73809d1a6f04f02",
                      "name": "circle 2*2",
                      "type": "2D",
                      "category": "676e9af6c73809d1a6f04f02",
                      "author": "authorId123",
                      "tags": [
                        "2d layer",
                        "component specific"
                      ],
                      "version": "1.0.1",
                      "svgFile": "circle.svg",
                      "svgContent": "<svg>...</svg>",
                      "diagram_components": [
                        {
                          "id": "shape-1736157286702-iynkldxxt",
                          "shape": "server-L",
                          "position": "top",
                          "source": "shape",
                          "relativeToId": null,
                          "attached2DShapes": [],
                          "attachmentPoints": [
                            {
                              "name": "attach-front-right",
                              "x": 226.2,
                              "y": 159.39999
                            },
                            {
                              "name": "attach-front-left",
                              "x": 63.700001,
                              "y": 159.39999
                            },
                            {
                              "name": "attach-back-left",
                              "x": 63.799999,
                              "y": 65.599998
                            },
                            {
                              "name": "attach-top",
                              "x": 144.89999,
                              "y": 76.400002
                            },
                            {
                              "name": "attach-back-right",
                              "x": 226.2,
                              "y": 65.599998
                            },
                            {
                              "name": "attach-bottom",
                              "x": 145,
                              "y": 148.60001
                            }
                          ]
                        },
                        {
                          "id": "shape-1736157286703-502hkdwrk",
                          "shape": "database-L",
                          "position": "top",
                          "source": "shape",
                          "relativeToId": "shape-1736157286702-iynkldxxt",
                          "attached2DShapes": [
                            {
                              "name": "react logo-L",
                              "attachedTo": "top"
                            }
                          ],
                          "attachmentPoints": [
                            {
                              "name": "attach-top",
                              "x": 106,
                              "y": 54
                            },
                            {
                              "name": "attach-bottom",
                              "x": 106,
                              "y": 126.2
                            },
                            {
                              "name": "attach-front-right",
                              "x": 187.2,
                              "y": 137
                            },
                            {
                              "name": "attach-back-right",
                              "x": 187.2,
                              "y": 43.200001
                            },
                            {
                              "name": "attach-front-left",
                              "x": 24.700001,
                              "y": 137
                            },
                            {
                              "name": "attach-back-left",
                              "x": 24.700001,
                              "y": 43.200001
                            }
                          ]
                        }
                      ],
                      "attachment_points": [
                        {
                          "name": "attach-top",
                          "x": 645,
                          "y": 504
                        },
                        {
                          "name": "attach-front-right",
                          "x": 726,
                          "y": 659
                        },
                        {
                          "name": "attach-front-left",
                          "x": 564,
                          "y": 659
                        },
                        {
                          "name": "attach-back-left",
                          "x": 564,
                          "y": 566
                        },
                        {
                          "name": "attach-back-right",
                          "x": 726,
                          "y": 566
                        },
                        {
                          "name": "attach-bottom",
                          "x": 645,
                          "y": 649
                        }
                      ]
                    }
                  ],
                  "total": 1
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters or filter format"
          },
          "404": {
            "description": "No shapes found matching the filter"
          }
        }
      }
    },
    "/shapes/category/{categoryId}": {
      "get": {
        "summary": "Get shapes by category ID",
        "tags": [
          "Shapes",
          "Categories"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "description": "The ID of the category to retrieve shapes for.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "677271fbc191b99b3f4b4d6d"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of shapes in the specified category",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Shape"
                      },
                      "example": [
                        {
                          "_id": "exampleShapeId1",
                          "name": "Shape 1",
                          "category": "677271fbc191b99b3f4b4d6d",
                          "description": "A description of shape 1",
                          "createdAt": "2024-12-30T11:15:05.294Z",
                          "updatedAt": "2024-12-30T11:15:05.294Z"
                        },
                        {
                          "_id": "exampleShapeId2",
                          "name": "Shape 2",
                          "category": "677271fbc191b99b3f4b4d6d",
                          "description": "A description of shape 2",
                          "createdAt": "2024-12-30T11:15:05.294Z",
                          "updatedAt": "2024-12-30T11:15:05.294Z"
                        }
                      ]
                    },
                    "total": {
                      "type": "integer",
                      "example": 2
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Category not found or incorrect category ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "incorrect category ID"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Shape": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the shape (required)"
          },
          "type": {
            "type": "string",
            "enum": [
              "2D",
              "3D",
              "LAYERS ",
              "COMPONENT"
            ],
            "description": "The type of the shape (required). Can be 2D, 3D, LAYERS, or COMPONENT."
          },
          "attachTo": {
            "type": "string",
            "description": "An optional field to attach the shape to another entity or category (optional)."
          },
          "svgFile": {
            "type": "string",
            "description": "An optional SVG file name or path (optional)"
          },
          "svgContent": {
            "type": "string",
            "description": "The content of the SVG (required for 2D and 3D shapes). It is a string representation of the SVG content."
          },
          "version": {
            "type": "string",
            "default": "1.0.0",
            "description": "The version of the shape (optional). This is useful to track different versions of the same shape."
          },
          "category": {
            "type": "string",
            "description": "A reference to the category to which the shape belongs (optional)."
          },
          "metadata": {
            "type": "object",
            "properties": {
              "description": {
                "type": "string",
                "description": "A detailed description of the shape (optional)."
              },
              "applicationTypes": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "Optional application types."
              },
              "customProperties": {
                "type": "object",
                "description": "Custom properties that allow for flexible, user-defined data (optional)."
              },
              "dependencies": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "shapeId": {
                      "type": "string"
                    },
                    "version": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "description": "Metadata for the shape, including description, application types, custom properties, and dependencies (optional)."
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "A list of tags associated with the shape, useful for categorization or filtering (optional)."
          },
          "author": {
            "type": "string",
            "description": "The author of the shape (required). This identifies who created the shape."
          },
          "diagram_components": {
            "type": "array",
            "description": "A list of components in the diagram associated with the shape. Each component can have various properties.",
            "items": {
              "type": "object",
              "description": "An arbitrary object representing a diagram component. The structure can vary for each component.",
              "additionalProperties": true
            }
          },
          "attachment_points": {
            "type": "array",
            "description": "A list of attachment points for the shape. Each attachment point can have various properties.",
            "items": {
              "type": "object",
              "description": "An arbitrary object representing an attachment point. The structure can vary for each attachment point.",
              "additionalProperties": true
            }
          }
        }
      }
    }
  }
}